edited map.h and second_part. changed roadLen type to unsigned int, now all functions that work with roads, are compatible with unsigned int road length. Bugfixed: 	1. Commands validation(), now roadLength are allowed only in unsigned int range. 	2. String_compareInt(), now doesn't overflow with big integers. 	3. String_toInt(), cleaned used safer implementation. Added: 	1. String_putInt(), makes string representation of given digit (int range int long).

+++ src/modules/mapNode.c
@@ -56,7 +56,7 @@ static MapNode *MapNode_getLessBranchedMapNode(MapNode *node1, MapNode *node2) {
     }
 }
 
-int min(int a, int b) {
+long int min(int a, int b) {
     if(a < b) {
         return a;
     }
@@ -95,7 +95,7 @@ Road *MapNode_getRoadFromConnectedNodes(MapNode *start, MapNode *target) {
     return NULL;
 }
 
-void MapNode_setDistanceFromRoot(MapNode *mapNode, int distance) {
+void MapNode_setDistanceFromRoot(MapNode *mapNode, unsigned distance) {
     mapNode->distanceFromRoot = distance;
 }
